bench: dump failing cycles' stamp streams in the measure loops
